requirements.txt
Flask==3.0.3
awsgi==0.0.5
gunicorn==22.0.0
psycopg2-binary
